+++ CHANGELOG.md
@@ -48,7 +48,8 @@ We could not reconstruct _all_ changes, but we tried our best to make the most o
 - [Installation] If there is no `config.php` file available during installation, create it during setup from the default config
 - [Discord] Introduced a new module to manage Discord Servers
 - [Party] Add information `Gesamt` in the Party box to show how many people can sign up for a party
-  
+- [Birthday] New module to show users birthdays
+
 ### Changed
 
 - [System] Raised the requirements of minimum PHP Version to v8.1